requirements.txt
six
google-api-python-client>=1.6.7
kaggle>=1.3.9
numpy==1.21.1
oauth2client
pandas==0.24.0
psutil>=5.4.3
py-cpuinfo>=3.3.0
split-folders
scikit-image
# dill>=0.3.4
keras==2.7.0
folium==0.2.1
# requests==2.24.0
pillow
lxml
jupyter
opencv-python
imageio
# ipykernel==4.10.1
ipython==5.5.0
ipython-genutils==0.2.0
ipython-sql==0.3.9


scipy>=0.19.1
tensorflow-text==2.7.3
tf-models-official==2.7.0
tensorflow==2.7.0
tensorflow-hub>=0.6.0
tensorflow-model-optimization>=0.4.1
tensorflow-datasets
tensorflow-addons
dataclasses;python_version<="3.7"
gin-config
tf_slim>=1.1.0
Cython
matplotlib
# Loader becomes a required positional argument in 6.0 in yaml.load
pyyaml>=5.1,<6.0
# google-colab
# CV related dependencies
opencv-python-headless
Pillow
pycocotools
# NLP related dependencies
seqeval
sentencepiece
sacrebleu

